feat: RESTFUL apis

+++ server/api/v1/presets/[id].delete.ts
@@ -1,5 +1,11 @@
 export default defineEventHandler(async (event) => {
   const id = getRouterParam(event, 'id')
+  if (!id) {
+    throw createError({
+      statusCode: 400,
+      message: 'id is required'
+    })
+  }
   const presetInfo = await prisma.preset.delete({
     where: {
       id
