+++ server/api/v1/templates/[id].delete.ts
@@ -1,5 +1,11 @@
 export default defineEventHandler(async (event) => {
   const id = getRouterParam(event, 'id')
+  if (!id) {
+    throw createError({
+      statusCode: 400,
+      message: 'id is required'
+    })
+  }
   const templateInfo = await prisma.template.delete({
     where: {
       id
